feat(step9): adversarial lock-before-snapshot-pin ordering regression
Stage 4's cross-process round-trip test proves the DSM handoff signal is
delivered correctly, but doesn't prove a copier that raced ahead of the
coordinator's table lock would actually be caught -- the online-snapshot
protocol's correctness rests entirely on the copier's first real SQL
(which pins its REPEATABLE READ snapshot) being unreachable before the
coordinator's table lock (plan §1e M2) is genuinely held.

New adversarial worker (external_zstd_handoff.rs): on waking from
LockHeldGoAhead, instead of trusting the signal, it independently
verifies the lock is real from its own separate session by attempting a
conflicting lock (ACCESS EXCLUSIVE NOWAIT, which conflicts with every
other lock mode including the coordinator's SHARE ROW EXCLUSIVE). If the
coordinator's lock is genuinely held this must fail with
lock_not_available; if it unexpectedly succeeds, the ordering guarantee
is broken and the test fails loudly instead of passing by coincidence.
Reuses HandoffPhase's existing terminal states (SnapshotPinned = conflict
correctly detected, Failed = ordering violation or unrelated error)
rather than adding a new shared-memory field. operation_nonce is
repurposed here to carry the coordinator's database oid so the worker can
connect to SPI in the right database -- documented as test-specific reuse
of that field, not a new protocol meaning.

Requires enable_spi_access (this worker does real SPI/lock work, unlike
the Stage 4 selftest worker) and the same crate-root #[unsafe(no_mangle)]
wrapper pattern (src/lib.rs) Stage 4 already established as necessary for
dynamic background worker function lookup.

test_copier_cannot_observe_go_ahead_before_lock_is_held (external_zstd_
handoff.rs): creates a real table, holds a real SHARE ROW EXCLUSIVE lock
across the coordinator's own signal exchange with a real second OS
process, and asserts the worker reports the conflict was correctly
detected.

Also: fixed two pre-existing clippy::unnecessary_cast lints in
external_zstd.rs (Stage 2), discovered because this is the first time
`cargo clippy` was run in this session with the pg_test feature enabled
(the #[cfg(test)]-gated code was never actually linted by the plain
`--features pg17` runs used throughout Stages 1-5) -- adopting
`--features pg17,pg_test` as part of the standard gate from here on.

cargo pgrx test pg17 (serial): 131/131 pass (130 existing + 1 new).

+++ src/storage/external_zstd.rs
@@ -463,7 +463,7 @@ mod tests {
     fn test_validate_root_os_level_accepts_good_root() {
         let dir = std::env::temp_dir().join(format!("fb-ext-zstd-test4-{}", std::process::id()));
         std::fs::create_dir_all(&dir).unwrap();
-        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(DIR_MODE as u32)).unwrap();
+        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(DIR_MODE)).unwrap();
         assert!(validate_root_os_level(dir.to_str().unwrap()).is_ok());
         std::fs::remove_dir_all(&dir).ok();
     }
@@ -472,7 +472,7 @@ mod tests {
     fn test_validate_root_os_level_rejects_symlink() {
         let dir = std::env::temp_dir().join(format!("fb-ext-zstd-test5-{}", std::process::id()));
         std::fs::create_dir_all(&dir).unwrap();
-        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(DIR_MODE as u32)).unwrap();
+        std::fs::set_permissions(&dir, std::fs::Permissions::from_mode(DIR_MODE)).unwrap();
         let link =
             std::env::temp_dir().join(format!("fb-ext-zstd-test5-link-{}", std::process::id()));
         let _ = std::fs::remove_file(&link);
